Remove separate worker file
This makes it easier to package this in e.g. esbuild.
